requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
rembg>=2.0.50
onnxruntime
tkinterdnd2
